Added stripe initializers. Nested cart and orders. Need to implement Orders Controller next. So that Orders controller can get the current cart id.
Need to generate the model for the Order with the attributes.
Order model with have FK reference to the cart id. So that the Order
model can get the total price from cart and calculate the final total
based on tax.

Cart total will become subtotal. Order will add the tax yielding new
total.

Charges controller will be able to get the order and the final
calculated total from the order object.